+++ packages/app/client/src/ui/dialogs/openBotDialog/openBotDialog.tsx
@@ -133,7 +133,7 @@ export class OpenBotDialog extends Component<OpenBotDialogProps, OpenBotDialogSt
 
     return (
       <Dialog cancel={this.props.onDialogCancel} className={openBotStyles.themeOverrides} title="Open a bot">
-        <form onSubmit={this.onSubmit}>
+        <form id="open-bot-dialog" onSubmit={this.onSubmit}>
           <div className={openBotStyles.autoCompleteBar}>
             <AutoComplete
               autoFocus={true}